requirements.txt
PyJWT==2.4.0
python-dotenv==0.20.0
cryptography==37.0.4
